+++ src/ckanapi_harvesters/builder/builder_resource_datastore_unmanaged.py
@@ -92,6 +92,7 @@ class BuilderDataStoreUnmanaged(BuilderDataStoreFile):  # , BuilderResourceUnman
                       resources_base_dir:str=None) -> CkanResourceInfo:
         """
         Specific implementation of patch_request which does not upload any data and only updates the fields currently present in the database
+
         :param resources_base_dir:
         :param ckan:
         :param package_id: