+++ src/ckanapi_harvesters/ckan_api/ckan_api_0_base.py
@@ -262,6 +262,7 @@ class CkanApiBase(CkanApiABC):
     def set_verbosity(self, verbosity:bool=True, verbose_extra:bool=None) -> None:
         """
         Enable/disable full verbose output
+
         :param verbosity: boolean. Cannot be None
         :return:
         """
@@ -370,7 +371,7 @@ class CkanApiBase(CkanApiABC):
 
         :param args:
         :param base_dir: base directory to find the CKAN API key file, if a relative path is provided
-        (recommended: leave None to use cwd)
+            (recommended: leave None to use cwd)
         :param error_not_found: option to raise an exception if the CKAN API key file is not found
         :param default_proxies: proxies used if proxies="default"
         :param proxy_headers: headers used to access the proxies, generally for authentication